Create a base class for broadlink entities
